Initial commit of the plugin.
The plugin unfreezes entities automatically when chunks are unloaded.

Additionally it has commands to unfreeze entities near players or in a certain area. It also has a command to freeze them, just in case.
The content of the README file is not written yet. Use "/etfu" in game to see the help menu.